fix(coordinator): repair quality gates and signal replay
