--- .npmrc
@@ -1,2 +1,0 @@
-onlyBuiltDependencies:
-  - esbuild
