+++ electron/mcp/preamble.ts
@@ -1,9 +1,9 @@
 import { randomUUID } from 'crypto';
 import { execFile } from 'child_process';
 import { promisify } from 'util';
-import { writeFileSync, readFileSync, existsSync, unlinkSync } from 'fs';
+import { readFileSync, existsSync, unlinkSync } from 'fs';
 import { readFile as fsReadFile, unlink as fsUnlink } from 'fs/promises';
-import { atomicWriteFile } from './atomic.js';
+import { atomicWriteFile, atomicWriteFileSync } from './atomic.js';
 import { join } from 'path';
 import os from 'os';
 
@@ -127,8 +127,8 @@ export async function buildNormalizedPreambleFileDiff(
   const tmpBase = join(os.tmpdir(), `parallel-code-base-${id}`);
   const tmpNorm = join(os.tmpdir(), `parallel-code-norm-${id}`);
   try {
-    writeFileSync(tmpBase, baseContent);
-    writeFileSync(tmpNorm, normalizedContent);
+    atomicWriteFileSync(tmpBase, baseContent);
+    atomicWriteFileSync(tmpNorm, normalizedContent);
     let diffOut = '';
     try {
       const { stdout } = await execAsync('git', ['diff', '--no-index', '-U3', tmpBase, tmpNorm]);
