+++ openspec/changes/archive/2026-05-30-custom-themes/proposal.md
