+++ package.json
@@ -21,7 +21,7 @@
     "lint:fix": "eslint . --fix",
     "lint:dead": "knip",
     "lint:arch": "depcruise --config .dependency-cruiser.cjs src electron",
-    "lint:security": "command -v semgrep >/dev/null 2>&1 && semgrep scan --config .semgrep/ --severity WARNING --error || (echo 'semgrep not installed (brew install semgrep or pip install semgrep)' >&2; exit 1)",
+    "lint:security": "sh -c 'command -v semgrep >/dev/null 2>&1 || { echo \"semgrep not installed (brew install semgrep or pip install semgrep)\" >&2; exit 1; }; semgrep scan --config .semgrep/ --severity WARNING --error'",
     "test:security-rules": "node scripts/test-semgrep-filesystem-safety.mjs",
     "lint:secrets": "command -v gitleaks >/dev/null 2>&1 && gitleaks detect --config .gitleaks.toml || (echo 'gitleaks not installed (brew install gitleaks)' >&2; exit 1)",
     "format": "prettier --write .",
@@ -59,7 +59,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.39.3",
-    "@types/dompurify": "^3.0.5",
+    "@fission-ai/openspec": "^1.3.1",
     "@types/node": "^25.3.0",
     "@types/qrcode": "^1.5.6",
     "@types/ws": "^8.18.1",
